Introduce itCount and keySize checks
By doing the first two checks in Java, we prevent lower level errors
(which would look severer to the end user, as shown in OPENJDK-991).

The third check is for consistency, we could just go ahead and return a
shorter key as SunJCE does (because of the integer floor-division by 8,
performed as 'keySize >> 3'), but that would later fail somewhere.