Use proper key types when deriving
When passing the CKA_KEY_TYPE attribute to the token, call
getPKCS11KeyType(), so we only get a valid PKCS#11 CKK_* constant.
This allows us to use the "HMAC" key type for HmacPBE* algorithms,
since getPKCS11KeyType("HMAC") returns CKK_GENERIC.

For the attributes' template manager, call getKeyType(), which keeps
SunPKCS11's 'pseudo' key types (PCKK_* constants, such as PCKK_HMAC).

For the Java key, keep the original algorithm of the SecretKeyFactory,
to align with SunJCE behaviour (this also keeps a trace of the algorithm
used to derive the key).