Review: fix P11PBEKey, rename PBES2Helper
Give a new serialVersionUID to P11PBEKey and return clones of salt and
password. Rename PBES2Helper to PBES2Params and move the DEFAULT_COUNT
and DEFAULT_SALT_LENGTH constants to it (both SunJCE and SunPKCS11 were
using the same values).

Also reset P11Mac in a single point of engineInit and fix some imports
order and whitespace inconsistencies.

Co-Authored-By: Martin Balao <mbalao@redhat.com>