Review: integrate KDFData into P11SecretKeyFactory
Move the P11Util.KDFData map to P11SecretKeyFactory, renaming KDFData to
KeyInfo and merging the map with the keyTypes map. KeyInfo has a
PBEKeyInfo specialization when the algorithm it represents is a PBE one.

Implement all the service and key algorithms consistency checks inside
P11SecretKeyFactory.KeyInfo.checkUse.

Store the PBEKeyInfo entry in advance for the service's algorithm in
P11Mac, P11SecretKeyFactory and P11PBECipher (when the algorithm is a
PBE one, which for P11PBECipher this means always).

Keep a single version of P11SecretKeyFactory.derivePBEKey, which now
receives a PBEKeyInfo instance, instead of the algorithm as a string.

Add some checks in P11SecretKeyFactory.convertKey and remove unnecessary
ones.

Partially undo 'Use proper key types when deriving' changes and pass the
native key type to Token.getAttributes, as P11SecretKeyFactory.createKey
does the same. We've also checked that the template system ignores the
passed SunPKCS11 pseudo key type when getting attributes.

Adjust all the remaining code for these changes and make some cosmetic
changes, such as using pattern matching for instanceof (JEP 305) or
reordering some bits in P11SecretKeyFactory.engineGenerateSecret and
P11SecretKeyFactory.engineGetKeySpec.

Co-Authored-By: Martin Balao <mbalao@redhat.com>